+++ adrilight/Spots/ILEDSetup.cs
@@ -18,8 +18,8 @@ namespace adrilight.Spots
         string Description { get; set; }
         int MatrixWidth { get; set; }
         int MatrixHeight { get; set; }
-        int PixelWidth { get; set; }
-        int PixelHeight { get; set; }
+        double PixelWidth { get; set; }
+        double PixelHeight { get; set; }
         object Lock { get; }
         int SetupID { get; set; }    // to match with device ID
         void DimLED(float dimFactor);
